requirements.txt
numpy
pandas==2.1.4
scikit-learn==1.2.2
xgboost==2.1.1
python-dotenv
fastapi
uvicorn[standard]